workers/requirements.txt
# Worker服务专用依赖包
# 基础依赖
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pika==1.3.2

# 图像处理和OCR
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
easyocr==1.7.0
paddlepaddle==2.5.2
paddleocr==2.7.0.3

# 对象存储
minio==7.2.0

# 自然语言处理
jieba==0.42.1
scikit-learn==1.3.2

# 其他工具
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
requests==2.31.0